Generating report improvement
The report will now print if all 4 scenarios are selected, this is because the Network visualization section still needs work. Perhaps using facets rather that using separate graphs per scenario. But I have not found a facet option in R for igraph package. 

Another issue to solve is that maps are not showing in the word document. But that does not stop generating the report.